+++ app/utils/design.ts
@@ -12,11 +12,6 @@ import {
 	type IDesignsByType,
 } from '#app/models/design.server'
 import { DesignTypeEnum, type designTypeEnum } from '#app/schema/design'
-import {
-	findFirstDesignIdInArray,
-	getNextDesignId,
-	getPrevDesignId,
-} from './artboard'
 
 export const filterAndOrderArtboardDesignsByType = ({
 	designs,
@@ -324,4 +319,40 @@ export const panelItemVariablesDesignType = ({
 		prevDesignId,
 		nextVisibleDesignId,
 	}
+}
+
+// helper for when a design is deleted, made invisible, or moved
+const getNextDesignId = (
+	visibleDesignIds: string[],
+	currentDesignId: string,
+): string | null => {
+	const currentIndex = visibleDesignIds.indexOf(currentDesignId)
+	if (currentIndex === -1) return null
+
+	return visibleDesignIds[currentIndex + 1] || null
+}
+
+// helper for when a design is made visible or moved
+const getPrevDesignId = (
+	visibleDesignIds: string[],
+	currentDesignId: string,
+): string | null => {
+	const currentIndex = visibleDesignIds.indexOf(currentDesignId)
+	if (currentIndex === -1) return null
+
+	return visibleDesignIds[currentIndex - 1] || null
+}
+
+const findFirstDesignIdInArray = (
+	designIds: string[],
+	designId1: string,
+	designId2: string,
+): string | null => {
+	const index1 = designIds.indexOf(designId1)
+	const index2 = designIds.indexOf(designId2)
+	if (index1 === -1 && index2 === -1) return null // Neither value is found
+	if (index1 !== -1 && index2 === -1) return designId1 // Only designId1 is found
+	if (index1 === -1 && index2 !== -1) return designId2 // Only designId2 is found
+	// Both values are found, return the one with the lower index
+	return index1 < index2 ? designId1 : designId2
 }
